pyproject.toml
[project]
name = "bpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb>=0.3.27",
    "datasets>=4.8.3",
    "docling>=2.90.0",
    "faiss-cpu>=1.7.4",
    "fsspec>=2026.2.0",
    "huggingface-hub>=1.7.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-groq>=1.1.2",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.2",
    "open-clip-torch>=3.3.0",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pymupdf>=1.27.2",
    "pypdf>=6.9.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
]
